Group funcionando.

+++ src/main/java/com/example/tfg/service/implementation/CourseServiceImp.java
@@ -1,5 +1,6 @@
 package com.example.tfg.service.implementation;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -14,6 +15,7 @@ import com.example.tfg.repository.CourseDao;
 import com.example.tfg.service.AcademicTermService;
 import com.example.tfg.service.ActivityService;
 import com.example.tfg.service.CourseService;
+import com.example.tfg.service.GroupService;
 
 @Service
 public class CourseServiceImp implements CourseService {
@@ -23,6 +25,9 @@ public class CourseServiceImp implements CourseService {
 
 	@Autowired
 	ActivityService serviceActivity;
+	
+	@Autowired
+	GroupService serviceGroup;
 
 	// @Autowired
 	// private DegreeService serviceDegree;
@@ -71,22 +76,17 @@ public class CourseServiceImp implements CourseService {
 	public List<Course> getAll() {
 		return daoCourse.getAll();
 	}
-//yo lo borraria no tiene sentido en este caso el modify
+
 	@Transactional(readOnly = false)
 	public boolean modifyCourse(Course course, Long id_academic, Long id_course) {
 		Course courseModify = daoCourse.getCourse(id_course);
 		
-		course.setAcademicTerm(serviceAcademicTerm.getAcademicTerm(id_academic));
-		Course existModify = daoCourse.exist(course);
+		courseModify.setSubject(course.getSubject());
 		
-		
-		if (existModify == null){
-				courseModify.setSubject(course.getSubject());
-				return daoCourse.saveCourse(courseModify);
-		}
+		return daoCourse.saveCourse(courseModify);
 
 			
-		else return false;
+
 	}
 
 	@Transactional(readOnly = false)
@@ -97,7 +97,10 @@ public class CourseServiceImp implements CourseService {
 	@Transactional(propagation = Propagation.REQUIRED)
 	public boolean deleteCourse(Long id) {
 		Course course = daoCourse.getCourse(id);
-		if (serviceActivity.deleteActivitiesFromCourse(course))
+		Collection <Course> coursesList = new ArrayList<Course>();
+		boolean deleteActivities = serviceActivity.deleteActivitiesFromCourse(course);
+		boolean deleteGroups = serviceGroup.deleteGroupsFromCourses(coursesList);
+		if(deleteActivities && deleteGroups)	
 			return daoCourse.deleteCourse(id);
 		return false;
 	}
@@ -120,9 +123,9 @@ public class CourseServiceImp implements CourseService {
 	}
 
 	public boolean deleteCoursesFromAcademic(AcademicTerm academic) {
-		boolean deleted = serviceActivity.deleteActivitiesFromCourses(academic.getCourses());
-		
-		if (deleted)
+		boolean deleteActivities = serviceActivity.deleteActivitiesFromCourses(academic.getCourses());
+		boolean deleteGroups = serviceGroup.deleteGroupsFromCourses(academic.getCourses());
+		if (deleteActivities && deleteGroups)
 			return daoCourse.deleteCoursesFromAcademic(academic);
 		else return false;
 //		boolean deleted = false;
@@ -139,10 +142,13 @@ public class CourseServiceImp implements CourseService {
 	
 
 
-	
+	@Transactional(readOnly = true)
 	public Course getCourseAll(Long id) {
 		Course c = daoCourse.getCourse(id);
 		c.setActivities(serviceActivity.getActivitiesForCourse(id));
+		c.setGroups(serviceGroup.getGroupsForCourse(id));
+		
+	
 		return c;
 	}
 
@@ -157,7 +163,8 @@ public class CourseServiceImp implements CourseService {
 	public boolean deleteCourses(Collection<AcademicTerm> academicList) {
 		Collection<Course> coursesList = daoCourse.getCoursesFromListAcademic(academicList);
 		boolean deleteActivities = serviceActivity.deleteActivitiesFromCourses(coursesList);
-		if (deleteActivities)
+		boolean deleteGroups = serviceGroup.deleteGroupsFromCourses(coursesList);
+		if (deleteActivities && deleteGroups)
 		return daoCourse.deleteCourses(academicList);
 		else return false;
 	}
